+++ MVVM/Models/MoneyBox.cs
@@ -4,5 +4,5 @@ public partial class MoneyBox
 {
     public int Id { get; set; }
 
-    public double Money { get; set; }
+    public int Money { get; set; }
 }
